+++ hoomd/md/TwoStepNVE.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2009-2019 The Regents of the University of Michigan
+// Copyright (c) 2009-2021 The Regents of the University of Michigan
 // This file is part of the HOOMD-blue project, released under the BSD 3-Clause License.
 
 
@@ -371,8 +371,8 @@ void TwoStepNVE::integrateStepTwo(unsigned int timestep)
 void export_TwoStepNVE(py::module& m)
     {
     py::class_<TwoStepNVE, IntegrationMethodTwoStep, std::shared_ptr<TwoStepNVE> >(m, "TwoStepNVE")
-        .def(py::init< std::shared_ptr<SystemDefinition>, 
-                       std::shared_ptr<ParticleGroup>, 
+        .def(py::init< std::shared_ptr<SystemDefinition>,
+                       std::shared_ptr<ParticleGroup>,
                        bool >())
         .def_property("limit", &TwoStepNVE::getLimit,
                                &TwoStepNVE::setLimit)